Initial refactor, moved common functionality into the Transport trait. four fn's to impl per transport.
