+++ src/sdmmc/mod.rs
@@ -1,0 +1,148 @@
+
+
+mod sdio;
+mod spi;
+
+
+use crate::sdmmc_proto::*;
+
+pub use spi::SdMmcSpi;
+
+pub trait Transport {
+
+    fn init(&mut self) -> Result<(), Error>;
+
+    fn card_command(&self, command: u8, arg: u32) -> Result<u8, Error>;
+
+    /// Receive a byte from the SD card
+    fn receive(&self) -> Result<u8, Error>;
+
+    /// Send a byte to the SD card.
+    fn send(&self, out: u8) -> Result<(), Error>;
+
+    /// Read an arbitrary number of bytes from the card. Always fills the
+    /// given buffer, so make sure it's the right size.
+    fn read_data(&self, buffer: &mut [u8]) -> Result<(), Error> {
+        // Get first non-FF byte.
+        let mut delay = Delay::new();
+        let status = loop {
+            let s = self.receive()?;
+            if s != 0xFF {
+                break s;
+            }
+            delay.delay(Error::TimeoutReadBuffer)?;
+        };
+        if status != DATA_START_BLOCK {
+            return Err(Error::ReadError);
+        }
+
+        for b in buffer.iter_mut() {
+            *b = self.receive()?;
+        }
+
+        let mut crc = u16::from(self.receive()?);
+        crc <<= 8;
+        crc |= u16::from(self.receive()?);
+
+        let calc_crc = crc16(buffer);
+        if crc != calc_crc {
+            return Err(Error::CrcError(crc, calc_crc));
+        }
+
+        Ok(())
+    }
+
+    /// Write an arbitrary number of bytes to the card.
+    fn write_data(&self, token: u8, buffer: &[u8]) -> Result<(), Error> {
+        let calc_crc = crc16(buffer);
+        self.send(token)?;
+        for &b in buffer.iter() {
+            self.send(b)?;
+        }
+        self.send((calc_crc >> 16) as u8)?;
+        self.send(calc_crc as u8)?;
+        let status = self.receive()?;
+        if (status & DATA_RES_MASK) != DATA_RES_ACCEPTED {
+            Err(Error::WriteError)
+        } else {
+            Ok(())
+        }
+    }
+    
+    fn card_acmd(&self, command: u8, arg: u32) -> Result<u8, Error> {
+        self.card_command(CMD55, 0)?;
+        self.card_command(command, arg)
+    }
+}
+
+const DEFAULT_DELAY_COUNT: u32 = 32_000;
+/// A terrible hack for busy-waiting the CPU while we wait for the card to
+/// sort itself out.
+///
+/// @TODO replace this!
+struct Delay(u32);
+impl Delay {
+    fn new() -> Delay {
+        Delay(DEFAULT_DELAY_COUNT)
+    }
+
+    fn delay(&mut self, err: Error) -> Result<(), Error> {
+        if self.0 == 0 {
+            Err(err)
+        } else {
+            let dummy_var: u32 = 0;
+            for _ in 0..100 {
+                unsafe { core::ptr::read_volatile(&dummy_var) };
+            }
+            self.0 -= 1;
+            Ok(())
+        }
+    }
+}
+
+/// The possible errors `SdMmcSpi` can generate.
+#[derive(Debug, Copy, Clone)]
+pub enum Error {
+    /// We got an error from the SPI peripheral
+    Transport,
+    /// We failed to enable CRC checking on the SD card
+    CantEnableCRC,
+    /// We didn't get a response when reading data from the card
+    TimeoutReadBuffer,
+    /// We didn't get a response when waiting for the card to not be busy
+    TimeoutWaitNotBusy,
+    /// We didn't get a response when executing this command
+    TimeoutCommand(u8),
+    /// We didn't get a response when executing this application-specific command
+    TimeoutACommand(u8),
+    /// We got a bad response from Command 58
+    Cmd58Error,
+    /// We failed to read the Card Specific Data register
+    RegisterReadError,
+    /// We got a CRC mismatch (card gave us, we calculated)
+    CrcError(u16, u16),
+    /// Error reading from the card
+    ReadError,
+    /// Error writing to the card
+    WriteError,
+    /// Can't perform this operation with the card in this state
+    BadState,
+    /// Couldn't find the card
+    CardNotFound,
+}
+
+/// The possible states `SdMmcSpi` can be in.
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum State {
+    NoInit,
+    Error,
+    Idle,
+}
+
+/// The different types of card we support.
+#[derive(Debug, Copy, Clone, PartialEq)]
+enum CardType {
+    SD1,
+    SD2,
+    SDHC,
+}
